refactor(pdf): extract line boundary analysis
